feat(reagent entity): change the type of quantity to double

+++ src/main/java/com/example/sennova/application/dto/inventory/ReagentInventory/UsageReagentRequest.java
@@ -6,7 +6,7 @@ public record UsageReagentRequest(
         String responsibleName,
         Long reagentId,
         @NotNull(message = "No se puede guardar un registro sin la cantidad")
-        Integer quantity,
+        Double quantity,
         String notes
 ) {
 }
